# recipes/zeebe-external-tasklist/tasklistApi/openapi.yaml
openapi: 3.0.1
info:
  title: Tasklist API
  version: 1.0.0
paths:
...
  /rest/task/load:
    post:
      summary: Load a task form and it's data
      operationId: loadData
      tags:
        - load-task-controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserTaskDto"
      responses:
        '200':
          description: Task form loaded
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/HtmlForm"
                  - $ref: "#/components/schemas/JsonForm"
...
components:
  schemas:
...
    FormData:
      type: object
    UserTaskDto:
      type: object
      required:
        - bpmnProcessId
        - elementId
        - key
        - processDefinitionKey
        - processInstanceKey
        - taskState
        - variables
      properties:
        key:
          type: integer
          format: int64
        elementId:
          type: string
        processInstanceKey:
          type: integer
          format: int64
        bpmnProcessId:
          type: string
        processDefinitionKey:
          type: integer
          format: int64
        variables:
          type: object
          additionalProperties:
            type: object
        taskState:
          type: string
        assignee:
          type: string
...
    FormDto:
      type: object
    HtmlForm:
      type: object
      required:
        - html
        - updatable
      allOf:
        - $ref: "#/components/schemas/FormDto"
        - type: object
          properties:
            html:
              type: string
            updatable:
              type: boolean
            formData:
              $ref: "#/components/schemas/FormData"
    JsonForm:
      type: object
      required:
        - schema
        - uiSchema
        - updatable
      allOf:
        - $ref: "#/components/schemas/FormDto"
        - type: object
          properties:
            schema:
              type: string
            uiSchema:
              type: string
            updatable:
              type: boolean
            formData:
              $ref: "#/components/schemas/FormData"