# recipes/zeebe-external-tasklist/tasklistApi/src/main/kotlin/io/miragon/zeebe/tm/tasklist/UserTaskDto.kt
package io.miragon.zeebe.tm.tasklist

data class UserTaskDto(
    val key: Long,
    val elementId: String,
    val processInstanceKey: Long,
    val bpmnProcessId: String,
    val processDefinitionKey: Long,
    val variables: Map<String, Any>,
    var taskState: String,
    var assignee: String?,
)